squishy: main: hooked up the SD Card socket and spend a while fighting placing test points on the bottom only to just give up on that.
Sadly we might need to cut a whole bunch of them, but I need to at least make sure the important ones can be accessed, ideally from the bottom of th board but with where the SD card socket is at the moment that's really unlikely, and that whole area in general is packed really tight so there isn't much room on top either.

Maybe we can use some of the connector pads on top as test points, which we need for the MCU and flash, those are the two really big important bits that need test points for automated programming in a fixture, but mounting the board upside down is not  ideal.